Settle one of the twelve signedness conflicts by measuring it
Twelve globals in the tree are declared with two spellings differing only in
signedness. Rather than park another one behind a comment, this measures a
pair: gCardGrid_bCursorColumn and gCardGrid_bCursorRow, declared s8 in
func_8002A788.c and u8-with-an-(s8)-cast in func_8002BFCC.c.

Editing only func_8002BFCC.c -- func_8002A788.c carries hand-written
R_MIPS_GPREL16 .reloc directives naming these symbols, so changing it would
confound the result -- gives four cases:

    extern u8  + (s8) cast     matches (what master had)
    extern s8  + (s8) cast     matches, object byte-identical
    extern s8  + no cast       matches, object byte-identical
    extern u8  + no cast       FAILS, executable 16 bytes short

The fourth is the control. Without it, three passes would be equally well
explained by the edit never reaching the build.

So the declaration is not load-bearing here; the requirement is that the read
be signed, and three spellings express that identically. GCC 2.8.1 emits
lbu/sll/sra even when told s8, because the shifted value is a shared
subexpression the following sra ...,0x1f reuses.

This conflict is therefore resolvable, and it resolves toward the spelling
that states the requirement once: both sources now say s8 and neither casts.
The note says explicitly not to generalize to the other eleven without
measuring them, since this pair had a consumer that was already casting and
a pair that reads directly is a different shape.

Verified with a clean `make match`: 84a54ed7...a45c88, and func_8002BFCC.o
is byte-identical to the object master produces.

+++ src/game/func_8002BFCC.c
@@ -16,8 +16,8 @@ extern s16 gGraphics_sViewportY_data asm("gGraphics_sViewportY")
     __attribute__((section(".data")));
 #define gGraphics_sViewportX gGraphics_sViewportX_data
 #define gGraphics_sViewportY gGraphics_sViewportY_data
-extern u8 gCardGrid_bCursorColumn;
-extern u8 gCardGrid_bCursorRow;
+extern s8 gCardGrid_bCursorColumn;
+extern s8 gCardGrid_bCursorRow;
 extern void (*D_800E9DBC[])(void);
 extern DuelEffectResourceRecord D_800EA0E8[];
 extern u8 D_800EA1E8[];
@@ -73,7 +73,7 @@ void func_8002BFCC(void) {
     func_80029590();
     r = D_800EA1E8;
     D_800EA1E8[0] = 0;
-    c = (s8)gCardGrid_bCursorColumn;
+    c = gCardGrid_bCursorColumn;
     d = c;
     if (c >= 0xA) {
         x = (c % 10) * 0xE + 0xAE;
@@ -82,7 +82,7 @@ void func_8002BFCC(void) {
     }
     *(s16 *)(r + 0x12) = x;
     *(s16 *)(r + 8) = x;
-    k = (s8)gCardGrid_bCursorRow;
+    k = gCardGrid_bCursorRow;
     y = (k / 10) * 0xB2 + (k % 10) * 0x10 + 0xE;
     *(s16 *)(r + 0x14) = y;
     *(s16 *)(r + 0xA) = y;
